+++ app/data/spells/air.ts
@@ -9,6 +9,7 @@ export const airSpell: Array<Spell> = [
     effect: "Target (or all) allied troop's speed is " +
       "increased by 3 (or 5) hexes per turn",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Haste.png`,
+    magic: "air",
   },
   {
     id: "35",
@@ -16,6 +17,7 @@ export const airSpell: Array<Spell> = [
     level: 1,
     effect: "Reveals location of all unclaimed artifacts",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/View_Air.png`,
+    magic: "air",
   },
   {
     id: "36",
@@ -24,6 +26,7 @@ export const airSpell: Array<Spell> = [
     effect: "Modifies the army composition scouting report " +
       "visible to enemies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Disguise.png`,
+    magic: "air",
   },
   {
     id: "37",
@@ -32,6 +35,7 @@ export const airSpell: Array<Spell> = [
     effect: "Reduces defense rating of a single enemy. " +
       "May be used repeatedly on the same target",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Disrupting_Ray.png`,
+    magic: "air",
   },
   {
     id: "38",
@@ -39,6 +43,7 @@ export const airSpell: Array<Spell> = [
     level: 2,
     effect: "Increases luck. Basic: Single ally. Expert: All allies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Fortune.png`,
+    magic: "air",
   },
   {
     id: "39",
@@ -46,6 +51,7 @@ export const airSpell: Array<Spell> = [
     level: 2,
     effect: "Inflicts magic damage on a single enemy",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Lightning_Bolt.png`,
+    magic: "air",
   },
   {
     id: "40",
@@ -54,6 +60,7 @@ export const airSpell: Array<Spell> = [
     effect: " 	Increases ranged attack rating. Basic: Single " +
       "ally. Expert: All allies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Precision.png`,
+    magic: "air",
   },
   {
     id: "41",
@@ -63,6 +70,7 @@ export const airSpell: Array<Spell> = [
       "Single ally. Expert: All allies. Effect increased " +
       "from 30%-30%-50%-50% to 50%-50%-75%-75%. Horn of the Abyss",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Protection_from_Air.png`,
+    magic: "air",
   },
   {
     id: "42",
@@ -70,6 +78,7 @@ export const airSpell: Array<Spell> = [
     level: 3,
     effect: "Reduces damage taken from ranged attacks",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Air_Shield.png`,
+    magic: "air",
   },
   {
     id: "43",
@@ -77,6 +86,7 @@ export const airSpell: Array<Spell> = [
     level: 3,
     effect: "Inflicts magic damage on all undead creatures",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Destroy_Undead.png`,
+    magic: "air",
   },
   {
     id: "44",
@@ -84,6 +94,7 @@ export const airSpell: Array<Spell> = [
     level: 3,
     effect: "Allows control of actions taken by enemy creature",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Hypnotize.png`,
+    magic: "air",
   },
   {
     id: "45",
@@ -92,6 +103,7 @@ export const airSpell: Array<Spell> = [
     effect: "Inflicts magic damage on initial creature and " +
       "reduced damage on nearby creatures",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Chain_Lightning.png`,
+    magic: "air",
   },
   {
     id: "46",
@@ -100,6 +112,7 @@ export const airSpell: Array<Spell> = [
     effect: " 	Allows extra retaliations per round. Basic: " +
       "Single ally. Expert: All allies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Counterstrike.png`,
+    magic: "air",
   },
   {
     id: "47",
@@ -107,6 +120,7 @@ export const airSpell: Array<Spell> = [
     level: 5,
     effect: "Teleports hero to visible location",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Dimension_Door.png`,
+    magic: "air",
   },
   {
     id: "48",
@@ -116,6 +130,7 @@ export const airSpell: Array<Spell> = [
       "all map obstacles. Cannot fly through cave " +
       "walls in the underground",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Fly.png`,
+    magic: "air",
   },
   {
     id: "49",
@@ -125,6 +140,7 @@ export const airSpell: Array<Spell> = [
       "20% chance of occurrence. Advanced: 30% chance. Expert: " +
       "40% chance",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Magic_Mirror.png`,
+    magic: "air",
   },
   {
     id: "50",
@@ -132,6 +148,7 @@ export const airSpell: Array<Spell> = [
     level: 5,
     effect: "Summons allied Air Elementals for the duration of combat",
     imageUrl: `${APP_DOMAIN}assets/images/spells/air/Summon_Air_Elemental.png`,
+    magic: "air",
   },
   {
     id: "51",
@@ -141,5 +158,6 @@ export const airSpell: Array<Spell> = [
       "Requires Titan's Thunder",
     imageUrl:
       `${APP_DOMAIN}assets/images/spells/air/Titan's_Lightning_Bolt.png`,
+    magic: "air",
   },
 ];