+++ app/data/spells/earth.ts
@@ -9,6 +9,7 @@ export const earthSpell: Array<Spell> = [
     effect: "Reduces damage taken from hand-to-hand attacks. " +
       "Basic: Single ally. Expert: All allies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Shield.png`,
+    magic: "earth",
   },
   {
     id: "53",
@@ -19,6 +20,7 @@ export const earthSpell: Array<Spell> = [
       "new reduced formula (old speed x 0.5 + 1). Horn of " +
       "the Abyss",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Slow.png`,
+    magic: "earth",
   },
   {
     id: "54",
@@ -27,6 +29,7 @@ export const earthSpell: Array<Spell> = [
     effect: "Increases defense rating. Basic: Single ally. " +
       "Expert: All allies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Stone_Skin.png`,
+    magic: "earth",
   },
   {
     id: "55",
@@ -34,6 +37,7 @@ export const earthSpell: Array<Spell> = [
     level: 1,
     effect: "Reveals location of all resources",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/View_Earth.png`,
+    magic: "earth",
   },
   {
     id: "56",
@@ -41,6 +45,7 @@ export const earthSpell: Array<Spell> = [
     level: 2,
     effect: "Inflicts magic damage on all living creatures",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Death_Ripple.png`,
+    magic: "earth",
   },
   {
     id: "57",
@@ -51,6 +56,7 @@ export const earthSpell: Array<Spell> = [
       "Creatures attempting to pass through quicksand become stuck " +
       "and make the location of the quicksand visible",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Quicksand.png`,
+    magic: "earth",
   },
   {
     id: "58",
@@ -59,6 +65,7 @@ export const earthSpell: Array<Spell> = [
     effect: "Reanimates killed undead creatures. Reanimated creatures " +
       "are not lost when combat ends",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Animate_Dead.png`,
+    magic: "earth",
   },
   {
     id: "59",
@@ -66,6 +73,7 @@ export const earthSpell: Array<Spell> = [
     level: 3,
     effect: "Grants spell immunity to a single creature",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Anti-Magic.png`,
+    magic: "earth",
   },
   {
     id: "60",
@@ -74,6 +82,7 @@ export const earthSpell: Array<Spell> = [
     effect: "Inflicts damage on random castle wall sections " +
       "during a siege",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Earthquake.png`,
+    magic: "earth",
   },
   {
     id: "61",
@@ -82,6 +91,7 @@ export const earthSpell: Array<Spell> = [
     effect: "Places force field at specified location. " +
       "Creatures cannot pass this field",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Force_Field.png`,
+    magic: "earth",
   },
   {
     id: "62",
@@ -92,6 +102,7 @@ export const earthSpell: Array<Spell> = [
       "from 30%-30%-50%-50% to 50%-50%-75%-75%. Horn of the Abyss",
     imageUrl:
       `${APP_DOMAIN}assets/images/spells/earth/Protection_from_Earth.png`,
+    magic: "earth",
   },
   {
     id: "63",
@@ -99,6 +110,7 @@ export const earthSpell: Array<Spell> = [
     level: 4,
     effect: "Inflicts magic damage on all creatures in area",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Meteor_Shower.png`,
+    magic: "earth",
   },
   {
     id: "64",
@@ -108,6 +120,7 @@ export const earthSpell: Array<Spell> = [
       "end of combat. Advanced: Resurrected creatures are " +
       "not lost when combat ends",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Resurrection.png`,
+    magic: "earth",
   },
   {
     id: "65",
@@ -115,6 +128,7 @@ export const earthSpell: Array<Spell> = [
     level: 4,
     effect: "Reduces morale. Basic: Single enemy. Expert: All enemies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Sorrow.png`,
+    magic: "earth",
   },
   {
     id: "66",
@@ -123,6 +137,7 @@ export const earthSpell: Array<Spell> = [
     effect: "Teleports hero to closest allied, unoccupied town. " +
       "Advanced: May choose any allied, unoccupied town",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Town_Portal.png`,
+    magic: "earth",
   },
   {
     id: "67",
@@ -130,6 +145,7 @@ export const earthSpell: Array<Spell> = [
     level: 5,
     effect: "Inflicts magic damage on a single enemy",
     imageUrl: `${APP_DOMAIN}assets/images/spells/earth/Implosion.png`,
+    magic: "earth",
   },
   {
     id: "68",
@@ -138,5 +154,6 @@ export const earthSpell: Array<Spell> = [
     effect: "Summons allied Earth Elementals for the duration of combat",
     imageUrl:
       `${APP_DOMAIN}assets/images/spells/earth/Summon_Earth_Elemental.png`,
+    magic: "earth",
   },
 ];