+++ app/data/spells/fire.ts
@@ -9,6 +9,7 @@ export const fireSpell: Array<Spell> = [
     effect: "Increases the attack skill of target (or all) " +
       "allied creature for melee attacks",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Bloodlust.png`,
+    magic: "fire",
   },
   {
     id: "2",
@@ -17,6 +18,7 @@ export const fireSpell: Array<Spell> = [
     effect: "The targeted (or all) enemy unit(s) deliver" +
       "minimum or reduced damage when attacking",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Curse.png`,
+    magic: "fire",
   },
   {
     id: "3",
@@ -27,6 +29,7 @@ export const fireSpell: Array<Spell> = [
       "increased from 30%-30%-50%-50% to 50%-50%-75%-75%. " +
       "Horn of the Abyss",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Protection_from_Fire.png`,
+    magic: "fire",
   },
   {
     id: "4",
@@ -35,6 +38,7 @@ export const fireSpell: Array<Spell> = [
     effect: "Prevents any action from being taken by " +
       "target enemy creature. ",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Blind.png`,
+    magic: "fire",
   },
   {
     id: "5",
@@ -44,6 +48,7 @@ export const fireSpell: Array<Spell> = [
       "that inflicts magic damage on creatures that" +
       "pass through",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Fire_Wall.png`,
+    magic: "fire",
   },
   {
     id: "6",
@@ -51,6 +56,7 @@ export const fireSpell: Array<Spell> = [
     level: 3,
     effect: "Inflicts magic damage on all creatures in area",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Fireball.png`,
+    magic: "fire",
   },
   {
     id: "7",
@@ -59,6 +65,7 @@ export const fireSpell: Array<Spell> = [
     effect: "locations on the battlefield. Only visible " +
       "to caster and creatures native to the terrain",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Land_Mine.png`,
+    magic: "fire",
   },
   {
     id: "8",
@@ -66,6 +73,7 @@ export const fireSpell: Array<Spell> = [
     level: 3,
     effect: "Reduces luck. Basic: Single enemy. Expert: All enemies",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Misfortune.png`,
+    magic: "fire",
   },
   {
     id: "9",
@@ -75,6 +83,7 @@ export const fireSpell: Array<Spell> = [
       "Damage reduced from 50*(spell power) + 30/30/60/120 to 40*(spell power) " +
       "+ 30/30/60/120. Horn of the Abyss",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Armageddon.png`,
+    magic: "fire",
   },
   {
     id: "10",
@@ -82,6 +91,7 @@ export const fireSpell: Array<Spell> = [
     level: 4,
     effect: "Forces enemy to attack closest creature",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Berserk.png`,
+    magic: "fire",
   },
   {
     id: "11",
@@ -90,6 +100,7 @@ export const fireSpell: Array<Spell> = [
     effect: "Inflicts magic damage proportional to " +
       "damage taken from hand-to-hand attacks",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Fire_Shield.png`,
+    magic: "fire",
   },
   {
     id: "12",
@@ -97,6 +108,7 @@ export const fireSpell: Array<Spell> = [
     level: 4,
     effect: "Increases attack rating and sets defense rating to 0",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Frenzy.png`,
+    magic: "fire",
   },
   {
     id: "13",
@@ -104,6 +116,7 @@ export const fireSpell: Array<Spell> = [
     level: 4,
     effect: "Inflicts magic damage on all creatures in area",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Inferno.png`,
+    magic: "fire",
   },
   {
     id: "14",
@@ -113,6 +126,7 @@ export const fireSpell: Array<Spell> = [
       "Dragons, Behemoths and Hydras. Advanced: Effect applies" +
       "to Angels and Devils. Expert: Effect applies to Titans",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Slayer.png`,
+    magic: "fire",
   },
   {
     id: "15",
@@ -121,6 +135,7 @@ export const fireSpell: Array<Spell> = [
     effect: "Destroys and removes living, unkilled allies to bring" +
       "previously living, killed allies back to life",
     imageUrl: `${APP_DOMAIN}assets/images/spells/fire/Sacrifice.png`,
+    magic: "fire",
   },
   {
     id: "16",
@@ -129,5 +144,6 @@ export const fireSpell: Array<Spell> = [
     effect: "Summons allied Fire Elementals for the duration of combat",
     imageUrl:
       `${APP_DOMAIN}assets/images/spells/fire/Summon_Fire_Elemental.png`,
+    magic: "fire",
   },
 ];