+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add specific endpoint for the town creatures.
 - Add middleware for the creatures endpoint.
+- Add specific endpoint for the magic spells.
 
 #v0.3.1